tests: move hostdisk to libstorage
We place tests in the storage package, but re-importable helper
functions in libstorage.
